fix: request tests corrected
